LPS-69488 Move Node.js default version configuration to gradle-plugins
This way, it will be applied in Workspaces, too.
