+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/internal/NodeDefaultsPlugin.java
@@ -16,6 +16,7 @@ package com.liferay.gradle.plugins.internal;
 
 import com.liferay.gradle.plugins.BaseDefaultsPlugin;
 import com.liferay.gradle.plugins.internal.util.GradleUtil;
+import com.liferay.gradle.plugins.node.NodeExtension;
 import com.liferay.gradle.plugins.node.NodePlugin;
 import com.liferay.gradle.plugins.node.tasks.ExecuteNodeTask;
 import com.liferay.gradle.plugins.node.tasks.ExecuteNpmTask;
@@ -37,6 +38,7 @@ public class NodeDefaultsPlugin extends BaseDefaultsPlugin<NodePlugin> {
 
 	@Override
 	protected void configureDefaults(Project project, NodePlugin nodePlugin) {
+		_configureNode(project);
 		_configureTasksExecuteNpm(project);
 		_configureTasksNpmInstall(project);
 		_configureTasksPublishNodeModule(project);
@@ -48,6 +50,13 @@ public class NodeDefaultsPlugin extends BaseDefaultsPlugin<NodePlugin> {
 	}
 
 	private NodeDefaultsPlugin() {
+	}
+
+	private void _configureNode(Project project) {
+		NodeExtension nodeExtension = GradleUtil.getExtension(
+			project, NodeExtension.class);
+
+		nodeExtension.setNodeVersion(_NODE_VERSION);
 	}
 
 	private void _configureTaskExecuteNpm(ExecuteNpmTask executeNpmTask) {
@@ -184,6 +193,8 @@ public class NodeDefaultsPlugin extends BaseDefaultsPlugin<NodePlugin> {
 
 		executeNodeTask.args(key + value);
 	}
+
+	private static final String _NODE_VERSION = "6.6.0";
 
 	private static final String _SASS_BINARY_SITE_ARG = "--sass-binary-site=";
